% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.z FUNCTIONS.R
\name{avg_month}
\alias{avg_month}
\title{Extract vertical water movements at the boundary between shallow and deep compartments}
\usage{
avg_month(data, grid_points)
}
\arguments{
\item{data}{A list of paths to netcdf files which share the same month.}

\item{grid_points}{An SF object containing the grid points which fall within the domain polygons.}
}
\value{
The function returns a dataframe of points averaged within a month and interpolated to the boundary depth.
}
\description{
This function reads in a packet of daily netcdf files from the same month, and returns a single monthly dataframe
of vertical water movements.
}
\details{
The function reads in a packet of monthly netcdf files and passes them to `avg_day`.

`avg_day` interpolates to the depth boundary between the shallow and deep compartments. These daily dataframes represent a spatial plane.

The daily dataframes are then passed back to `avg_month`, which averages again by month and variable, dropping the spatial information.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{decadal}()},
\code{\link{strip_ice}()},
\code{\link{summarise_sp}()},
\code{\link{summarise_ts_detritus}()},
\code{\link{summarise_ts}()}
}
\concept{NEMO-MEDUSA averages}
