% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NM.z FUNCTIONS.R
\name{Big_Currents}
\alias{Big_Currents}
\title{Extract a full field for sampling currents}
\usage{
Big_Currents(data)
}
\arguments{
\item{data}{A dataframe containing a batch of netcdf file locations which share a month and year, but may have different
variables.}
}
\value{
A dataframe containing points and lat/lon coordinates. Attached data are the average NEMO-MEDUSA model outputs for
zonal and meridional currents in the shallow and deep zone. The dataframe contains the data averaged across a whole month.
}
\description{
\strong{THIS FUNCTION IS LIKELY REDUNDANT}
}
\details{
Since this function was written the domain polygons have been reshaped, and I learnt more than just currents needed to be
sampled at the boundary of compartments. The more general `whole_month` now behaves in the same way, returning a grid clipped
to the window of interest for the project, but not the model domain.

The function takes a packet of netcdf files which contain model outputs for the same month. These are split by the type of
data they contain, and are then passed to `type_in_month` for variable specific averaging. After, the data are returned to this
function to be combined into a single monthly dataframe for all variables. The dataframe is saved in the Objects directory
under Currents.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{detritus_month}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_sea}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_month}()},
\code{\link{interp_vertical}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
