% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{box_zoom}
\alias{box_zoom}
\title{Zoom a ggplot to the bounding box of an sf object}
\usage{
box_zoom(sf)
}
\arguments{
\item{sf}{an sf object that you want to crop the plot window to}
}
\value{
Add to a ggplot with the standard `+` syntax to adjust the plotting window.
}
\description{
This function can be used with ggplot and uses an sf object to adjust the plotting window according to a bounding box.
If you want more breathing space use st_buffer when providing `sf`.
}
