% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary functions.R
\name{characterise_flows}
\alias{characterise_flows}
\title{Check if a segment runs parallel to latitude or longitude}
\usage{
characterise_flows(lines, domain, precision = 1000)
}
\arguments{
\item{lines}{An sfc object of lines representing transects along the perimeter of a model domain polygon.}

\item{domain}{An sfc polygon object of the model domain, used to make the transects.}

\item{precision}{How far apart should sampling points be (1/x degrees, larger numbers bring the points closer).}

\item{segment}{An SF line object representing a transect along the perimeter of a model domain polygon.}
}
\value{
The function returns a dataframe detailing whether laitudues are the same and longitudes are the same along a segment.

The function returns a dataframe detailing which two model compartments are either side of the transect,
and the direction of the exchange.

The function returns the lines object with added columns. If transects are ambiguosly labelled, a list is returned of the transects and the failed sampling locations.
}
\description{
This function checks whether the transects created to sample the boundaries of model compartments lie on
a lat-lon grid. If they do, then we know that when sampling meridional and zonal currents only one is needed
per transect.
}
\details{
The function takes a boundary segment (transect) and reprojects it on a lat-lon grid. The coordinates are extracted,
and checked for identical latitudes or longitudes.
}
\seealso{
Other Boundary sampling functions: 
\code{\link{boundaries}()},
\code{\link{to_segments}()}
}
\concept{Boundary sampling functions}
\concept{Boundary sampling functions
Characterise transects (weights, target current, nature of water exchanges)
This function takes a transect from the edge of a model compartment and works out where it is spatially. This is
neccessary for correctly averaging variables along the whole boundary of a compartment.

The function determines which model compartment is next to the focal compartment. The focal compartment is either the
offshore or inshore polygon, with the offshore polygon selected ahead of the inshore polygon when these share a boundary.
This defines the link between model compartments.

Direction (in/out) is also calculated relative to the focal polygon. At the western boundary of a polygon, positive zonal
currents \emph{(West to East)} indicate water flowing into the model compartment. However, at the Eastern boundary,
flows from West to East now leave the model compartment. A True or False column called flip is included. If a positive value
for a water velocity is leaving the polygon, or a negative flow value is entering the polygon, flip can be used to multiply
by -1 to correctly sum the water movements from the perspective of the focal polygon.}
\concept{Boundary sampling functions
Test whether "positive" currents flow in or out of a model compartment at a segment

This function takes a transect from the edge of a model compartment and works out where it is spatially. This is
neccessary for correctly averaging variables along the whole boundary of a compartment.

The function determines which model compartment is next to the focal compartment. The focal compartment is either the
offshore or inshore polygon, with the offshore polygon selected ahead of the inshore polygon when these share a boundary.
This defines the link between model compartments.

Direction (in/out) is also calculated relative to the focal polygon. At the western boundary of a polygon, positive zonal
currents \emph{(West to East)} indicate water flowing into the model compartment. However, at the Eastern boundary,
flows from West to East now leave the model compartment. A True or False column called flip is included. If a positive value
for a water velocity is leaving the polygon, or a negative flow value is entering the polygon, flip can be used to multiply
by -1 to correctly sum the water movements from the perspective of the focal polygon.}
