% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit conversion functions.R
\name{Convert-units}
\alias{Convert-units}
\alias{micro_to_full}
\alias{full_to_micro}
\alias{micro_to_milli}
\alias{milli_to_micro}
\alias{milli_to_full}
\alias{full_to_milli}
\alias{l_to_m3}
\alias{m3_to_l}
\alias{sec_to_day}
\alias{day_to_sec}
\title{Helper functions for common unit conversions}
\usage{
micro_to_full(data)

full_to_micro(data)

micro_to_milli(data)

milli_to_micro(data)

milli_to_full(data)

full_to_milli(data)

l_to_m3(data)

m3_to_l(data)

sec_to_day(data)

day_to_sec(data)
}
\arguments{
\item{data}{A numeric vector.}
}
\value{
A numeric vector following unit conversion.
}
\description{
These functions are named to make it intuitive to switch between scales, with pairs existing for
different directions of conversion.
}
\details{
Though some of these functions do the same thing numerically, having multiple names which match the desired
conversion should reduce friction when developing code. This should also help avoid typos when it comes to
0s, and also stop you needing to constantly look up conversion factors!
}
