% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{execute}
\alias{execute}
\title{Run an R Script in it's own Session Informatively}
\usage{
execute(filepath)
}
\arguments{
\item{filepath}{a character string indicating the R script you want to execute.}
}
\value{
The R script is run for it's side effects and the run time added to a log by `tictoc`.
}
\description{
This function is a wrapper for a few tweaks around sourcing an R script.
}
\details{
The R script is run in it's own R session, to avoid crashing an R session handling the batch processing of scripts.
Activity is indicated with a spinner, and completed scripts are announced to help localise errors to a script. The time
taken for a script to complete is also tracked in a log file, which can be accessed with `tictoc::tic.log()`.
}
