% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfw.z FUNCTIONS.R
\name{FAO_boats}
\alias{FAO_boats}
\title{Extract the values from a grid under transects along the external boundaries of the model domain}
\usage{
FAO_boats(Data, Clip)
}
\arguments{
\item{Data}{The data object as provided by Sample_OOB}

\item{Depth}{The depth layer to extract data from. Either "S" or "D"}

\item{variables}{The variables to extract, provided by Sample_OOB}
}
\value{
The function returns a dataframe of transects and their average DIN, chlorophyll, temperature, and salinity values by depth.
}
\description{
This function reads in a datafile and attaches the values needed to transects.
It uses a precalculated set of indices of where transects intersect the grid for speed.
}
