% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{fill_in}
\alias{fill_in}
\title{Fill missing values in a vector}
\usage{
fill_in(data)
}
\arguments{
\item{data}{a vector containing "" or NA as missing values.}
}
\value{
A filled in vector where "" or NA is replaced with the latest non-missing value.
}
\description{
This function takes a vector containing missing values ("" or NA), and overwrites empty positions with
the previous non-missing entry. This assumes the first value is not a missing value.
}
\details{
The motivating example for this was to automatically fill in cells in BODC nutrient data where metadata
is not copied past the first row of a block of data representing a single CTD cast.
}
\examples{
# Create dummy data with missing entries
test <- c(1, 2, "", "", NA, 3, "")

# Replace blanks
filled <- fill_in(test)
}
