% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEDIMENT functions.R
\name{get_bed_shear_stress}
\alias{get_bed_shear_stress}
\title{Calculate bed shear stress at a location}
\usage{
get_bed_shear_stress(pixel, tide_packets, depth = 40)
}
\value{
A dataframe containing a time series of bed shear stress metrics for a location.
}
\description{
This function takes tidal data from SINMOD and wave data from ECMWF, aligns the two time series at a single location,
and then calculates the bed shear stress through time.
}
\concept{water movements}
