% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFW functions.R
\name{get_cropped_fishing}
\alias{get_cropped_fishing}
\title{Import Fishing Effort Within Specified Areas}
\usage{
get_cropped_fishing(file, area)
}
\arguments{
\item{file}{The full name to a .csv file containing Global Fishing Watch data.}

\item{area}{A Simple Feature object representing the areas to return fishing activity in.}
}
\value{
A dataframe of cells from a 0.01 degree grid which contain fishing activity,
labelled by the FAO region they fall in.
}
\description{
Import Fishing Effort Within Specified Areas
}
\seealso{
Other Global Fishing Watch functions: 
\code{\link{FAO_boats}()},
\code{\link{jacks_lazy_animation}()},
\code{\link{jacks_lazy_avg}()},
\code{\link{jacks_polar_animate}()},
\code{\link{jacks_polar_static}()}
}
\concept{Global Fishing Watch functions}
