% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NM.z FUNCTIONS.R
\name{get_ice}
\alias{get_ice}
\title{Extract the values from a grid under transects along the external boundaries of the model domain}
\usage{
get_ice(path, file)
}
\arguments{
\item{Depth}{The depth layer to extract data from. Either "S" or "D"}

\item{Data}{The data object as provided by Sample_OOB}

\item{variables}{The variables to extract, provided by Sample_OOB}
}
\value{
The function returns a dataframe of transects and their average DIN, chlorophyll, temperature, and salinity values by depth.
}
\description{
This function reads in a datafile and attaches the values needed to transects.
It uses a precalculated set of indices of where transects intersect the grid for speed.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{Big_Currents}()},
\code{\link{avg_day}()},
\code{\link{detritus_month}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_merid}()},
\code{\link{get_sea}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
