% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions.R
\name{get_SPM}
\alias{get_SPM}
\title{Extract Suspended Particulate Matter data from Globcolour}
\usage{
get_SPM(File, Year, Month, crop)
}
\arguments{
\item{File}{a path to a GLobcolour netcdf file.}

\item{Year}{The year the file contains data for.}

\item{Month}{The month the file contains data for.}

\item{cropr}{The bounding box of an SF object to crop to, as returned by `st_bbox()`.}
}
\value{
A dataframe containing point estimates within an area of SPM with time columns appended.
}
\description{
This function takes a netcdf from Globcolour and returns the mean Suspended Particulate MAtter (SPM) from
within a lat/lon window.
}
