% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEDIMENT functions.R
\name{get_tides_ts}
\alias{get_tides_ts}
\title{Calculate bed shear stress at a location}
\usage{
get_tides_ts(File, month, vars, pixel_x, pixel_y)
}
\arguments{
\item{pixel}{a numeric value indicating the row number of a dataframe of target coordinates.}

\item{tide_packets}{a list of dataframes containing instructions for processing all SINMOD files which share a year.}
}
\value{
A dataframe containing a time series of bed shear stress metrics for a location.
}
\description{
This function takes tidal data from SINMOD and wave data from ECMWF, aligns the two time series at a single location,
and then calculates the bed shear stress through time.
}
\seealso{
Other water movements: 
\code{\link{get_bed_shear_stress}()},
\code{\link{get_waves_ts}()},
\code{\link{reshape_SINMOD}()},
\code{\link{sample_tides_year}()},
\code{\link{sample_waves}()}
}
\concept{water movements}
