% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary functions.R
\name{interp_month}
\alias{interp_month}
\title{Extract vertical water movements at the boundary between shallow and deep compartments}
\usage{
interp_month(data, grid_points)
}
\arguments{
\item{data}{A list of paths to netcdf files which share the same month.}

\item{grid_points}{An SF object containing the grid points which fall within the domain polygons.}
}
\value{
The function returns a dataframe of points averaged within a month and interpolated to the boundary depth.
}
\description{
This function reads in a packet of daily netcdf files from the same month, and returns a single monthly dataframe
of vertical water movements.
}
\details{
The function reads in a packet of monthly netcdf files and passes them to `interp_vertical`.

`interp_vertical` interpolates to the depth boundary between the shallow and deep compartments. These daily dataframes represent a spatial plane.

The daily dataframes are then passed back to `interp_month`, which averages again by month and variable, dropping the spatial information.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{detritus_month}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_sea}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_vertical}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
