% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boundary functions.R
\name{interp_vertical}
\alias{interp_vertical}
\title{Interpolate vertical water movements to the boundary depth between shallow and deep compartments}
\usage{
interp_vertical(Path, File, Points, Boundary = 60)
}
\arguments{
\item{Path}{The path to a netcdf file containing target data, inherited from `interp_month`.}

\item{File}{The name of a netcdf file containing target data, inherited from `interp_month`.}

\item{Points}{An SF object containing the grid points which fall within the domain polygons, inherited from `interp_month`.}

\item{Boundary}{The boundary depth between shallow and deep compartments, defaults to 60 m for MiMeMo.}
}
\value{
A data frame containing vertical eddy diffusivity and vertical velocity at the boundary depth
between shallow and deep compartments, for a single day as a spatial plane.
}
\description{
This function reads in a 3 dimensional array of water movements, and uses the `rcdo` package to interpolate
values to a plane at the boundary depth between shallow and deep compartments.
}
\details{
The function uses `rcdo::nc_remap` to interpolate vertical eddy diffusivity and vertical water velocities at
the boundary depth between the shallow and deep model compartments. Points which fall outside the geographic
extent of model compartments are dropped. The retained points are passed back to `interp_month` for further averaging
as a dataframe.
}
\seealso{
Other NEMO-MEDUSA variable extractors: 
\code{\link{detritus_month}()},
\code{\link{get_air_dt}()},
\code{\link{get_air}()},
\code{\link{get_bio}()},
\code{\link{get_detritus}()},
\code{\link{get_ice}()},
\code{\link{get_merid}()},
\code{\link{get_sea}()},
\code{\link{get_spatial}()},
\code{\link{get_vertical}()},
\code{\link{get_zonal}()},
\code{\link{interp_month}()},
\code{\link{type_in_month}()},
\code{\link{whole_month}()}
}
\concept{NEMO-MEDUSA variable extractors}
