% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiMeMo.tools-package.R
\docType{package}
\name{MiMeMo.tools-package}
\alias{MiMeMo.tools}
\alias{MiMeMo.tools-package}
\title{MiMeMo.tools: Tools for extracting data for StrathE2E, during MiMeMo}
\description{
Intended for use with the StarthE2Epolar model. These tools can clip datasets to the geographical area of interest. Data can then be summarised as averages to parameterise volumes and boundary conditions needed by StarthE2Epolar.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jack-h-laverick.github.io/MiMeMo.tools/}
  \item \url{https:github.com/Jack-H-Laverick/MiMeMo.tools}
  \item Report bugs at \url{https:github.com/Jack-H-Laverick/MiMeMo.tools/issues}
}

}
\author{
\strong{Maintainer}: Jack H Laverick \email{jacklaverick@ymail.com} (\href{https://orcid.org/0000-0001-8829-2084}{ORCID})

}
\keyword{internal}
