% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEDIMENT functions.R
\name{mud_to_permeability}
\alias{mud_to_permeability}
\title{Calculate sediment permeability}
\usage{
mud_to_permeability(percent_mud, scalar = -2.171, constant = -10.232)
}
\arguments{
\item{percent_mud}{The amount of all sediment at a location smaller than sand (mud/silt and finer) in percent.}

\item{scalar}{A parameter in the relationship}

\item{constant}{A parameter in the relationship.}
}
\value{
A numeric vector containing estimates of sediment permeability in m^2 (can be length 1).
}
\description{
This function takes a percent of mud/silt at a location and returns a measure of permeability. The default
parameterisation comes from Matt Pace's thesis for fine sediments.
}
\examples{
data <- data.frame(Silt = seq(0.5, 100, length.out = 100)) \%>\%
mutate(Permeability = mud_to_permeability(Silt))

ggplot(data) +
geom_line(aes(x = Silt, y = Permeability)) +
scale_y_continuous(trans = "log", breaks = c(0.000000000000001, 0.0000000000001, 0.00000000001))
}
\seealso{
Other Sediment properties: 
\code{\link{D50_to_porosity}()}
}
\concept{Sediment properties}
