% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NM functions.R
\name{NM_decadal_summary}
\alias{NM_decadal_summary}
\title{Summarise NEMO-MEDUSA Output into Decadal Grids
This function averages cleaned NEMO-MEDUSA monthly summaries into decadal grids.}
\usage{
NM_decadal_summary(decade, dt)
}
\arguments{
\item{dt}{Switch for using either data.table or dplyr methods (TRUE/FALSE respectively)}

\item{saved}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs. It must contain the columns:
Longitude, Latitude, Decade, Month, Shore, and Depth.}
}
\value{
A dataframe containing a summarised decade of spatialy resolved NEMO-MEDUSA outputs.
}
\description{
The function groups by all spatial variables (Longitude, Latitude, Depth, and Shore zone), and by decade and month.
The mean for every other variable is calculated within these groups.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{NM_boundary_summary}()},
\code{\link{NM_volume_summary}()},
\code{\link{decadal}()},
\code{\link{strip_ice}()}
}
\concept{NEMO-MEDUSA averages}
