% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NM functions.R
\name{NM_volume_summary}
\alias{NM_volume_summary}
\title{Summarise NEMO-MEDUSA Output into Time Series Within Model Compartments
This function averages NEMO-MEDUSA monthly summaries into time series for each model compartment.}
\usage{
NM_volume_summary(saved)
}
\arguments{
\item{saved}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs.}
}
\value{
A dataframe containing a mean monthly time series of all target variables in NEMO-MEDUSA outputs.
}
\description{
The function groups by model compartment (Depth and Shore zone) and time step (Month and Year).
The mean for every target variable is calculated within these groups.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{NM_boundary_summary}()},
\code{\link{NM_decadal_summary}()},
\code{\link{decadal}()},
\code{\link{strip_ice}()}
}
\concept{NEMO-MEDUSA averages}
