% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extractor functions other.R
\name{reproj}
\alias{reproj}
\title{Reproject from Latitude and Longitude to Project CRS}
\usage{
reproj(data, crs)
}
\arguments{
\item{data}{A dataframe containing Longitude and Latitude.}

\item{crs}{The new Coordinate Reference System  to project to.}
}
\value{
A dataframe, now with an x and y column specifying the coordinates for points in the projects Coordiante Reference System.
}
\description{
This function takes a dataframe containing a latitude and longitude column, and replaces them with an X and Y column of coordinates
in a new CRS.
}
\details{
The function converts a dataframe into an SF object and reprojects into a new CRS. Two coordinate columns are extracted from the
geometry column using `sfc_as_cols`, before the geometry column is dropped.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{Window}()},
\code{\link{sfc_as_cols}()},
\code{\link{vectors_2_direction}()}
}
\concept{NEMO-MEDUSA spatial tools}
