% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.z FUNCTIONS.R
\name{Sample}
\alias{Sample}
\title{Extract water flows across transects along the boundaries of model compartments}
\usage{
Sample(file)
}
\arguments{
\item{file}{Path to the .rds object containing both Zonal and Meridional water velocities.}
}
\value{
The function returns a dataframe of net water movements between model compartments
and the external environment in a given month.
}
\description{
This function calculates the water exchanges between model compartments and external boundaries.
}
\details{
A datafile containing water velocities is read in and split by depth before calling the extract function to
pull the correct velocities.

Sample then groups exchanges by depth, shore zone, neighbour and direction. Flows
are summed to return net water movements for the given time step.
}
