% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.z FUNCTIONS.R
\name{Sample_OOB}
\alias{Sample_OOB}
\title{Sample along the external boundaries of the model domain}
\usage{
Sample_OOB(file)
}
\arguments{
\item{file}{Path to the .rds object containing data.}
}
\value{
The function returns a dataframe of average DIN, chlorophyll, temperature, and salinity at
the external model boundary by compartment for a month.
}
\description{
This function is a variant of Sample which only considers the external boundaries of the model domain for
variables other than water movement.
}
\details{
A datafile is imported and split by depth before calling extract_OOB to sample the grid.

After extraction, Sample_OOB calculates the average for variables of interest by depth and shore zone.
}
