% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{stratify}
\alias{stratify}
\title{Extract the values from a grid under transects along the external boundaries of the model domain}
\usage{
stratify(data, depth, weights)
}
\arguments{
\item{Depth}{The depth layer to extract data from. Either "S" or "D"}

\item{Data}{The data object as provided by Sample_OOB}

\item{variables}{The variables to extract, provided by Sample_OOB}
}
\value{
The function returns a dataframe of transects and their average DIN, chlorophyll, temperature, and salinity values by depth.
}
\description{
This function reads in a datafile and attaches the values needed to transects.
It uses a precalculated set of indices of where transects intersect the grid for speed.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{Compartmentalise}()},
\code{\link{Window}()},
\code{\link{empty}()},
\code{\link{get_spatial}()},
\code{\link{get_weights.W}()},
\code{\link{get_weights.old}()},
\code{\link{get_weights}()},
\code{\link{reproj}()},
\code{\link{sfc_as_cols}()}
}
\concept{NEMO-MEDUSA spatial tools}
