% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{strip_ice}
\alias{strip_ice}
\title{Extract the values from a grid under transects along the external boundaries of the model domain}
\usage{
strip_ice(data)
}
\arguments{
\item{Depth}{The depth layer to extract data from. Either "S" or "D"}

\item{Data}{The data object as provided by Sample_OOB}

\item{variables}{The variables to extract, provided by Sample_OOB}
}
\value{
The function returns a dataframe of transects and their average DIN, chlorophyll, temperature, and salinity values by depth.
}
\description{
This function reads in a datafile and attaches the values needed to transects.
It uses a precalculated set of indices of where transects intersect the grid for speed.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{decadal}()},
\code{\link{summarise_sp}()},
\code{\link{summarise_ts_detritus}()},
\code{\link{summarise_ts}()}
}
\concept{NEMO-MEDUSA averages}
