% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEMO-MEDUSA functions.R
\name{summarise_sp_dt}
\alias{summarise_sp_dt}
\title{Average into Decadal Grids}
\usage{
summarise_sp_dt(decade)
}
\arguments{
\item{saved}{A dataframe containing a summarised month from NEMO-MEDUSA model outputs. It must contain the columns:
Longitude, Latitude, Decade, Month, Shore, and Depth.}
}
\value{
A dataframe containing a summarised decade of spatialy resolved NEMO-MEDUSA outputs.
}
\description{
This function averages cleaned NEMO-MEDUSA monthly summaries into decadal grids.
}
\details{
The function groups by all spatial variables (Longitude, Latitude, Depth, and Shore zone), and by decade and month.
The mean for every other variable is calculated within these groups.
}
\seealso{
Other NEMO-MEDUSA averages: 
\code{\link{decadal}()},
\code{\link{strip_ice_dt}()},
\code{\link{strip_ice}()},
\code{\link{summarise_sp}()},
\code{\link{summarise_ts_detritus}()},
\code{\link{summarise_ts}()}
}
\concept{NEMO-MEDUSA averages}
