% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{Update-drivers}
\alias{Update-drivers}
\alias{update_boundary_period}
\alias{update_physics_period}
\alias{update_physics_period_MA}
\title{Reproject from Latitude and Longitude to Project CRS}
\usage{
update_boundary_period(start, end, path, convert_units = TRUE)

update_physics_period(start, end, path)

update_physics_period_MA(start, end, path)
}
\arguments{
\item{start}{The first year in the target time period.}

\item{path}{A character string containing the path to a model variant, i.e. "./Models/Region/start-last".}

\item{convert_units}{logical for backwards compatibility with MiMeMo scripts. Should be set to `FALSE` and unit conversion performed before saving .rds files.}

\item{data}{A dataframe containing Longitude and Latitude.}

\item{crs}{The new Coordinate Reference System  to project to.}

\item{last}{The last year in the target time period.}
}
\value{
A dataframe, now with an x and y column specifying the coordinates for points in the projects Coordiante Reference System.

Run this function for it's side effects. The appropriate driving file will be updated.
}
\description{
This function takes a dataframe containing a latitude and longitude column, and replaces them with an X and Y column of coordinates
in a new CRS.
}
\details{
The function converts a dataframe into an SF object and reprojects into a new CRS. Two coordinate columns are extracted from the
geometry column using `sfc_as_cols`, before the geometry column is dropped.
}
\seealso{
Other NEMO-MEDUSA spatial tools: 
\code{\link{Window}()},
\code{\link{sfc_as_cols}()},
\code{\link{vectors_2_direction}()}
}
\concept{NEMO-MEDUSA spatial tools}
