% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global Fishing Watch functions.R
\name{FAO_boats}
\alias{FAO_boats}
\title{Extract the values from a grid under transects along the external boundaries of the model domain}
\usage{
FAO_boats(Data, Clip)
}
\arguments{
\item{Data}{The data object as provided by Sample_OOB}

\item{Depth}{The depth layer to extract data from. Either "S" or "D"}

\item{variables}{The variables to extract, provided by Sample_OOB}
}
\value{
The function returns a dataframe of transects and their average DIN, chlorophyll, temperature, and salinity values by depth.
}
\description{
This function reads in a datafile and attaches the values needed to transects.
It uses a precalculated set of indices of where transects intersect the grid for speed.
}
\seealso{
Other Global Fishing Watch functions: 
\code{\link{get_cropped_fishing}()},
\code{\link{jacks_lazy_animation}()},
\code{\link{jacks_lazy_avg}()},
\code{\link{jacks_polar_animate}()},
\code{\link{jacks_polar_static}()}
}
\concept{Global Fishing Watch functions}
