% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc functions.R
\name{Update-drivers}
\alias{Update-drivers}
\alias{update_boundary_period}
\alias{update_physics_period}
\title{Re-parameterise a Driver File for a Time Period in the MiMeMo Domain}
\usage{
update_boundary_period(start, end, path)

update_physics_period(start, end, path)
}
\arguments{
\item{start}{The first year in the target time period.}

\item{path}{A character string containing the path to a model variant, i.e. "./Models/Region/start-last".}

\item{last}{The last year in the target time period.}
}
\value{
Run this function for it's side effects. The appropriate driving file will be updated.
}
\description{
These function take a StrathE2E model and update the parameters in the driving data files to a specific time window.
}
\details{
The functions are informative. Warnings are produced to highlight which parameters don't have data for the target time
window. In this case the old parameters are left in place. The function will refuse to update parameters based on data
which covers less than half of the target time interval. For parameters which are updated, the function returns the
years used.
}
